# unison_openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/send-otp":{"post":{"operationId":"AuthController_sendOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendOtpDto"}}}},"responses":{"201":{"description":""}},"summary":"Send OTP to email for verification or forgot password","tags":["Authentication"]}},"/api/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":""}},"summary":"Verify OTP and receive a short-lived verified_token","tags":["Authentication"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"summary":"Register a new alumni or student account","tags":["Authentication"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"summary":"Login and receive a JWT token","tags":["Authentication"]}},"/api/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"summary":"Reset password using verified_token from OTP step","tags":["Authentication"]}},"/api/admin/pending-accounts":{"get":{"operationId":"AdminController_getPendingAccounts","parameters":[],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/approve-account/{id}":{"patch":{"operationId":"AdminController_approveAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApproveAccountDto"}}}},"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/reject-account/{id}":{"patch":{"operationId":"AdminController_rejectAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RejectAccountDto"}}}},"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/dashboard-stats":{"get":{"operationId":"AdminController_getDashboardStats","parameters":[],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/all-alumni":{"get":{"operationId":"AdminController_getAllAlumni","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/all-students":{"get":{"operationId":"AdminController_getAllStudents","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/admin/remove-account/{id}":{"delete":{"operationId":"AdminController_removeAccount","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RemoveAccountDto"}}}},"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/alumni/profile/{id}":{"get":{"operationId":"AlumniController_getProfile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Alumni"]},"put":{"operationId":"AlumniController_updateProfile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAlumniProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["Alumni"]}},"/api/alumni/work-experience":{"post":{"operationId":"AlumniController_addWorkExperience","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWorkExperienceDto"}}}},"responses":{"201":{"description":""}},"tags":["Alumni"]}},"/api/alumni/work-experience/{id}":{"put":{"operationId":"AlumniController_updateWorkExperience","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWorkExperienceDto"}}}},"responses":{"200":{"description":""}},"tags":["Alumni"]},"delete":{"operationId":"AlumniController_deleteWorkExperience","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Alumni"]}},"/api/alumni/skills":{"post":{"operationId":"AlumniController_addSkill","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddSkillDto"}}}},"responses":{"201":{"description":""}},"tags":["Alumni"]}},"/api/alumni/skills/{skill_id}":{"delete":{"operationId":"AlumniController_deleteSkill","parameters":[{"name":"skill_id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Alumni"]}},"/api/alumni/network":{"get":{"operationId":"AlumniController_getNetwork","parameters":[],"responses":{"200":{"description":""}},"tags":["Alumni"]}},"/api/alumni/connect/{target_id}":{"post":{"operationId":"AlumniController_connectWith","parameters":[{"name":"target_id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectDto"}}}},"responses":{"201":{"description":""}},"tags":["Alumni"]}},"/api/alumni/batch-mates":{"get":{"operationId":"AlumniController_getBatchMates","parameters":[],"responses":{"200":{"description":""}},"tags":["Alumni"]}},"/api/student/profile/{id}":{"get":{"operationId":"StudentController_getProfile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Student"]},"put":{"operationId":"StudentController_updateProfile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStudentProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["Student"]}},"/api/student/skills":{"post":{"operationId":"StudentController_addSkill","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddStudentSkillDto"}}}},"responses":{"201":{"description":""}},"tags":["Student"]}},"/api/student/mentors":{"get":{"operationId":"StudentController_getMentors","parameters":[],"responses":{"200":{"description":""}},"tags":["Student"]}},"/api/opportunities":{"post":{"operationId":"OpportunityController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOpportunityDto"}}}},"responses":{"201":{"description":""}},"tags":["Opportunity"]},"get":{"operationId":"OpportunityController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}},{"name":"skill","required":true,"in":"query","schema":{"type":"string"}},{"name":"is_remote","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Opportunity"]}},"/api/opportunities/my-posts":{"get":{"operationId":"OpportunityController_findMyPosts","parameters":[],"responses":{"200":{"description":""}},"tags":["Opportunity"]}},"/api/opportunities/{id}":{"get":{"operationId":"OpportunityController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Opportunity"]},"put":{"operationId":"OpportunityController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOpportunityDto"}}}},"responses":{"200":{"description":""}},"tags":["Opportunity"]},"delete":{"operationId":"OpportunityController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Opportunity"]}},"/api/search/alumni":{"get":{"operationId":"SearchController_searchAlumni","parameters":[{"name":"name","required":true,"in":"query","schema":{"type":"string"}},{"name":"company","required":true,"in":"query","schema":{"type":"string"}},{"name":"skill","required":true,"in":"query","schema":{"type":"string"}},{"name":"batch_year","required":true,"in":"query","schema":{"type":"string"}},{"name":"degree","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Search"]}},"/api/search/opportunities":{"get":{"operationId":"SearchController_searchOpportunities","parameters":[{"name":"title","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}},{"name":"skill","required":true,"in":"query","schema":{"type":"string"}},{"name":"location","required":true,"in":"query","schema":{"type":"string"}},{"name":"is_remote","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Search"]}},"/api/skills/all":{"get":{"operationId":"SkillController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Skill"]}},"/api/network/centrality":{"get":{"operationId":"NetworkController_getCentrality","parameters":[],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/network/shortest-path/{from_id}/{to_id}":{"get":{"operationId":"NetworkController_getShortestPath","parameters":[{"name":"from_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"to_id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/network/top-companies":{"get":{"operationId":"NetworkController_getTopCompanies","parameters":[],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/network/skill-trends":{"get":{"operationId":"NetworkController_getSkillTrends","parameters":[],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/network/batch-analysis":{"get":{"operationId":"NetworkController_getBatchAnalysis","parameters":[],"responses":{"200":{"description":""}},"tags":["Network"]}},"/api/notifications":{"get":{"operationId":"NotificationController_getNotifications","parameters":[],"responses":{"200":{"description":""}},"tags":["Notification"]}},"/api/notifications/{id}/read":{"patch":{"operationId":"NotificationController_markAsRead","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notification"]}}},"info":{"title":"UNISON API","description":"Alumni Network — CS Department, UET Faisalabad","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SendOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"ahmed@uet.edu.pk"},"type":{"type":"string","enum":["email_verification","forgot_password"]}},"required":["email","type"]},"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"ahmed@uet.edu.pk"},"otp":{"type":"string","example":"482910"},"type":{"type":"string","enum":["email_verification","forgot_password"]}},"required":["email","otp","type"]},"RegisterDto":{"type":"object","properties":{"verified_token":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["alumni","student"]},"roll_number":{"type":"string"},"degree":{"type":"string"},"graduation_year":{"type":"number","description":"Alumni only"},"semester":{"type":"number","description":"Student only"}},"required":["verified_token","name","email","password","role","roll_number","degree"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"ResetPasswordDto":{"type":"object","properties":{"verified_token":{"type":"string"},"new_password":{"type":"string"}},"required":["verified_token","new_password"]},"ApproveAccountDto":{"type":"object","properties":{}},"RejectAccountDto":{"type":"object","properties":{}},"RemoveAccountDto":{"type":"object","properties":{}},"UpdateAlumniProfileDto":{"type":"object","properties":{}},"CreateWorkExperienceDto":{"type":"object","properties":{}},"UpdateWorkExperienceDto":{"type":"object","properties":{}},"AddSkillDto":{"type":"object","properties":{}},"ConnectDto":{"type":"object","properties":{}},"UpdateStudentProfileDto":{"type":"object","properties":{}},"AddStudentSkillDto":{"type":"object","properties":{}},"CreateOpportunityDto":{"type":"object","properties":{}},"UpdateOpportunityDto":{"type":"object","properties":{}}}}}